backend/requirements.txt
PyMuPDF
python-dotenv
SQLAlchemy
pgvector
langchain-openai
# langchain_google_genai
# langchain_community
# deepinfra
langchain
langchain_experimental
psycopg2-binary
openai
# pyopencl
# torch
numpy
